# API Interface/notification-center-swagger (20-01-2025).yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Notifacation Center API",
    "description": "API documentation for Notifacation Center service",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8091",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/assignments/{assignmentId}": {
      "get": {
        "tags": [
          "assignment-controller"
        ],
        "summary": "Retrieve an Assignment given its ID",
        "operationId": "getAssignmentById",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assignment retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseAssignmentDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "401": {
            "description": "Authentication process failed!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "403": {
            "description": "Invalid authorization parameters. Check JWT or CSRF Token",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "417": {
            "description": "Assignment with id [ID] not found in DB",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerToken": []
          }
        ]
      },
      "put": {
        "tags": [
          "assignment-controller"
        ],
        "summary": "Update an assignment",
        "operationId": "updateAssignment",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Assignment updated successfully!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "401": {
            "description": "Authentication process failed!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "403": {
            "description": "Invalid authorization parameters. Check JWT or CSRF Token",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "417": {
            "description": "Expectation Failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "500": {
            "description": "Unable to create and store assignment",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerToken": []
          }
        ]
      },
      "delete": {
        "tags": [
          "assignment-controller"
        ],
        "summary": "Delete an assignment by ID",
        "operationId": "deleteAssignmentById_1",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assignment deleted successfully!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "401": {
            "description": "Authentication process failed!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "403": {
            "description": "Invalid authorization parameters. Check JWT or CSRF Token",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "417": {
            "description": "Assignment with id [ID] not found in DB",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerToken": []
          }
        ]
      }
    },
    "/api/assignments/{assignmentId}/comments": {
      "put": {
        "tags": [
          "assignment-controller"
        ],
        "summary": "Add a comment to an assignment depending on the user (Source or Target)",
        "operationId": "updateAssignmentComments",
        "parameters": [
          {
            "name": "assignmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentCommentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Assignment comments updated successfully!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "401": {
            "description": "Authentication process failed!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "403": {
            "description": "Invalid authorization parameters. Check JWT or CSRF Token",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "417": {
            "description": "Expectation Failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "500": {
            "description": "Unable to create and store assignment",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerToken": []
          }
        ]
      }
    },
...
    "/api/assignments/create": {
      "post": {
        "tags": [
          "assignment-controller"
        ],
        "summary": "Create a new assignment",
        "operationId": "createAssignment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Assignment created successfully!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "401": {
            "description": "Invalid JWT token inserted",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "403": {
            "description": "Invalid authorization parameters. Check JWT or CSRF Token",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "417": {
            "description": "Expectation Failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "500": {
            "description": "Unable to create and store assignment",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerToken": []
          }
        ]
      }
    },
...
    "/api/assignments": {
      "get": {
        "tags": [
          "assignment-controller"
        ],
        "summary": "Retrieve all Assignments",
        "operationId": "getAllAssignments",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sortAttribute",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "timestampUpdated"
            }
          },
          {
            "name": "isAscending",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Assignments retrieved successfully!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponsePaginatedResultsDtoAssignmentDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid sort attribute.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "401": {
            "description": "Authentication process failed!",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "403": {
            "description": "Invalid authorization parameters. Check JWT or CSRF Token",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseMapStringString"
                }
              }
            }
          },
          "417": {
            "description": "Expectation Failed",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseAppResponseString"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "BaseAppResponseString": {
        "type": "object",
        "properties": {
          "data": {
            "type": "string"
          },
          "errors": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BaseAppResponseMapStringString": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "errors": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AssignmentDto": {
        "type": "object",
        "properties": {
          "assignmentId": {
            "type": "string"
          },
          "sourceUserId": {
            "type": "string"
          },
          "targetUserId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "Open",
              "Accepted",
              "In Progress",
              "Completed"
            ]
          },
          "sourceUserComments": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "targetUserComments": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "description": {
            "type": "string"
          },
          "productionModule": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "timestampUpdated": {
            "type": "string",
            "format": "date-time"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ]
          }
        }
      },
      "AssignmentCommentDto": {
        "type": "object",
        "properties": {
          "sourceUserComment": {
            "type": "string"
          },
          "targetUserComment": {
            "type": "string"
          }
        }
      },
...
      "BaseAppResponsePaginatedResultsDtoAssignmentDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PaginatedResultsDtoAssignmentDto"
          },
          "errors": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PaginatedResultsDtoAssignmentDto": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssignmentDto"
            }
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int32"
          },
          "lastPage": {
            "type": "boolean"
          }
        }
      },
      "BaseAppResponseAssignmentDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AssignmentDto"
          },
          "errors": {
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}